refactor(npm-tools): rename jspDLS.js to getJSPTokenMatchers.js
